Move JsHandler into class file
